pyproject.toml
[tool.poetry]
name = "poc-cdc"
version = "0.1.0"
description = ""
authors = ["Wesley Oliveira <wesleyumsocaminho@gmail.com>"]
readme = "README.md"
packages = [{include = "poc_cdc"}]

[tool.poetry.dependencies]
python = "3.10.*"
pydantic = "^2.6.4"
requests = "^2.31.0"
sqlalchemy = "^2.0.29"
logging = "^0.4.9.6"
confluent-kafka = "^2.3.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
taskipy = "^1.12.2"
blue = "^0.9.1"
ruff = "^0.3.4"
httpx = "^0.27.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 79
exclude = ['.venv', 'migrations']

[tool.isort]
profile = "black"
line_length = 79
extend_skip = ['migrations']

[tool.pytest.ini_options]
pythonpath = "."

[tool.blue]
extend-exclude = '(migrations/)'


[tool.taskipy.tasks]
lint = 'ruff check . && blue --check . --diff && isort --check . --diff'
format = 'blue .  && isort .'
run = 'uvicorn fast_zero.app:app --reload'
pre_test = 'task lint'
test = 'pytest -s -x --cov=poc_cdc -vv'
post_test = 'coverage html'